Live plot: Remove duplication of NodeStatusMonitor
