+++ tools/studio/node_discovery.py
@@ -3,30 +3,11 @@ import datetime
 import sys
 import threading
 import time
+
 import uavcan
 
 from network.UavcanNode import UavcanNode
-
-class NodeStatusModel:
-    def __init__(self, node):
-        self.known_nodes = {}
-        self.node = node
-        self.node.add_handler(uavcan.protocol.NodeStatus, self._node_status_callback)
-
-    def _node_status_callback(self, event):
-        node_id = event.transfer.source_node_id
-        if node_id not in self.known_nodes:
-            self.known_nodes[node_id] = {}
-            self.node.request(uavcan.protocol.GetNodeInfo.Request(), node_id, self._response_callback)
-        self.known_nodes[node_id]['status'] = event.message
-
-    def _response_callback(self, event):
-        if not event:
-            raise RuntimeError("Remote call timeout")
-
-        board = event.transfer.source_node_id
-        name = str(event.response.name)
-        self.known_nodes[board]['name'] = name
+from network.NodeStatusMonitor import NodeStatusMonitor
 
 class NodeStatusViewer:
     def __init__(self):
@@ -98,7 +79,7 @@ def main():
 
     node = UavcanNode(interface=args.interface, node_id=args.node_id)
 
-    model = NodeStatusModel(node)
+    model = NodeStatusMonitor(node)
     viewer = NodeStatusViewer()
     controller = NodeStatusController(model, viewer)
