+++ tools/studio/node_selector.py
@@ -3,12 +3,14 @@ import datetime
 import sys
 import threading
 import time
-import uavcan
 
+import uavcan
 from pyqtgraph.Qt import QtCore, QtGui
 
 from network.UavcanNode import UavcanNode
+from network.NodeStatusMonitor import NodeStatusMonitor
 from viewers.Selector import SelectorWidget
+
 
 def parse_args():
     parser = argparse.ArgumentParser(description=__doc__)
@@ -17,34 +19,6 @@ def parse_args():
     parser.add_argument("--node_id", "-n", help="UAVCAN Node ID", default=127)
 
     return parser.parse_args()
-
-class NodeStatusMonitor:
-    def __init__(self, node):
-        self.known_nodes = {}
-        self.node = node
-        self.node.add_handler(uavcan.protocol.NodeStatus, self._node_status_callback)
-        self.on_new_node = None
-
-    def set_on_new_node_callback(self, on_new_node):
-        self.on_new_node = on_new_node
-
-    def _node_status_callback(self, event):
-        node_id = event.transfer.source_node_id
-        if node_id not in self.known_nodes:
-            self.known_nodes[node_id] = {}
-            self.node.request(uavcan.protocol.GetNodeInfo.Request(), node_id, self._response_callback)
-        self.known_nodes[node_id]['status'] = event.message
-
-    def _response_callback(self, event):
-        if not event:
-            raise RuntimeError("Remote call timeout")
-
-        board = event.transfer.source_node_id
-        name = str(event.response.name)
-        self.known_nodes[board]['name'] = name
-
-        if self.on_new_node is not None:
-            self.on_new_node()
 
 class NodeSelectController:
     def __init__(self, model, viewer):
